Overrode takeDamage in units to account for armor; Cast damage as int in takeDamage in entity as well.

+++ src/main/game/entities/Entity.java
@@ -28,7 +28,7 @@ public abstract class Entity {
     public double getCurrentHealth(){ return health; }                              // Return entity health
     public Percentage getHealthPercentage(){return healthPercent; }                 // Return entity health pct.
     public void takeDamage(double damage){                                          // Take damage to health
-        this.health -= damage;
+        this.health -= (int)damage;
         this.healthPercent.updateHealthPercentage((double)this.health);
     }
     public void heal(double healing){                                               // Heal for a given amount
